fix balance parsing

+++ AFTests/BlockchainsIntegration/BalancesTests.cs
@@ -160,8 +160,8 @@ namespace AFTests.BlockchainsIntegrationTests
                 sw.Start();
                 while (sw.Elapsed < TimeSpan.FromMinutes(10))
                 {
-                    if (int.Parse(request.Balances.GetBalances("500", null).GetResponseObject().Items.ToList().Find(a => a.Address == WALLET_SINGLE_USE).Balance) <
-                        int.Parse(startBalance))
+                    if (long.Parse(request.Balances.GetBalances("500", null).GetResponseObject().Items.ToList().Find(a => a.Address == WALLET_SINGLE_USE).Balance) <
+                        long.Parse(startBalance))
                     {
                         time = DateTime.Now.Ticks;
                         sw.Stop();
